chore: 댓글 아바타 출력하도록 Props 정의
