+++ apps/blog/src/app/posts/[slug]/_components/Comments/Comment.util.ts
@@ -9,5 +9,13 @@ export function convertToCommentProps(comment: NotionPage): CommentProps {
     createdAt: comment.properties.CreatedAt.created_time,
     liked: comment.properties.Liked.checkbox,
     owner: comment.properties.Owner.checkbox,
+    avatar: comment.properties.Avatar.url,
   };
 }
+
+export const getAvatarUrl = () => {
+  const styles = ['adventurer', 'avataaars', 'bottts', 'personas'];
+  const randomStyle = styles[Math.floor(Math.random() * styles.length)];
+  const randomSeed = Math.random().toString(36).substring(7);
+  return `https://api.dicebear.com/6.x/${randomStyle}/svg?seed=${randomSeed}`;
+};
